# Conflicts:
#	torchrl/data/tensor_specs.py
#	torchrl/envs/common.py

+++ torchrl/data/tensordict/memmap.py
@@ -8,12 +8,12 @@ from __future__ import annotations
 import functools
 import os
 import tempfile
-from math import prod
-from typing import Any, Callable, List, Optional, Tuple, Union
+from typing import Any, Callable, List, Optional, Tuple, Union, Dict
 
 import numpy as np
 import torch
 
+from torchrl import prod
 from torchrl.data.tensordict.utils import _getitem_batch_size
 from torchrl.data.utils import (
     DEVICE_TYPING,
@@ -24,6 +24,9 @@ from torchrl.data.utils import (
 MEMMAP_HANDLED_FN = {}
 
 __all__ = ["MemmapTensor", "set_transfer_ownership"]
+
+NoneType = type(None)
+EllipsisType = type(Ellipsis)
 
 
 def implements_for_memmap(torch_function) -> Callable:
@@ -52,7 +55,7 @@ class MemmapTensor(object):
     and remote workers that have access to
     a common storage, and as such it supports serialization and
     deserialization. It is possible to choose if the ownership is
-    transferred upon serialization / deserialization: If owenership is not
+    transferred upon serialization / deserialization: If ownership is not
     transferred (transfer_ownership=False, default), then the process where
     the MemmapTensor was created will be responsible of clearing it once it
     gets out of scope (in that process). Otherwise, the process that
@@ -105,6 +108,8 @@ class MemmapTensor(object):
         >>> assert not os.path.isfile(filename)
 
     """
+
+    requires_grad = False
 
     def __init__(
         self,
@@ -253,15 +258,27 @@ class MemmapTensor(object):
         self,
         idx: Optional[int] = None,
         memmap_array: Optional[np.ndarray] = None,
+        from_numpy: bool = False,
     ) -> torch.Tensor:
         if memmap_array is None:
             memmap_array = self.memmap_array
         if idx is not None:
+            if isinstance(idx, torch.Tensor):
+                idx = idx.cpu()
+            elif isinstance(idx, tuple) and any(
+                isinstance(sub_index, torch.Tensor) for sub_index in idx
+            ):
+                idx = tuple(
+                    sub_index.cpu()
+                    if isinstance(sub_index, torch.Tensor)
+                    else sub_index
+                    for sub_index in idx
+                )
             memmap_array = memmap_array[idx]
-        out = self._np_to_tensor(memmap_array)
+        out = self._np_to_tensor(memmap_array, from_numpy=from_numpy)
         if (
             idx is not None
-            and not isinstance(idx, (int, np.integer))
+            and not isinstance(idx, (int, np.integer, slice))
             and len(idx) == 1
             and not (isinstance(idx, torch.Tensor) and idx.dtype is torch.bool)
         ):  # and isinstance(idx, torch.Tensor) and len(idx) == 1:
@@ -269,7 +286,9 @@ class MemmapTensor(object):
             out = out.view(size)
         return out
 
-    def _np_to_tensor(self, memmap_array: np.ndarray) -> torch.Tensor:
+    def _np_to_tensor(self, memmap_array: np.ndarray, from_numpy: bool) -> torch.Tensor:
+        if from_numpy:
+            return torch.from_numpy(memmap_array)
         return torch.as_tensor(memmap_array, device=self.device)
 
     @classmethod
@@ -293,6 +312,11 @@ class MemmapTensor(object):
     def _tensor(self) -> torch.Tensor:
         return self._load_item()
 
+    @property
+    def _tensor_from_numpy(self) -> torch.Tensor:
+        # a tensor created with `from_numpy` to make sure that changes are done in-place
+        return self._load_item(from_numpy=True)
+
     def ndimension(self) -> int:
         return self._ndim
 
@@ -300,13 +324,13 @@ class MemmapTensor(object):
         return self._numel
 
     def clone(self) -> MemmapTensor:
-        """Clones the MemmapTensor onto another MemmapTensor
+        """Clones the MemmapTensor onto another tensor
 
         Returns:
-            a new MemmapTensor with the same data but a new storage.
+            a new torch.Tensor with the same data but a new storage.
 
         """
-        return MemmapTensor(self)
+        return self._tensor.clone()
 
     def contiguous(self) -> torch.Tensor:
         """Copies the MemmapTensor onto a torch.Tensor object.
@@ -382,10 +406,8 @@ class MemmapTensor(object):
         return self
 
     def __del__(self) -> None:
-        # if hasattr(self, "filename"):
-        if self._has_ownership:
+        if "_has_ownership" in self.__dir__() and self._has_ownership:
             os.unlink(self.filename)
-            # self.file.close()
 
     def __eq__(self, other: Any) -> torch.Tensor:
         if not isinstance(other, (MemmapTensor, torch.Tensor, float, int, np.ndarray)):
@@ -398,7 +420,9 @@ class MemmapTensor(object):
                 attr
             )  # make sure that appropriate exceptions are raised
 
-        if attr not in self.__getattribute__("_tensor_dir"):
+        if ("_tensor_dir" not in self.__dir__()) or (
+            attr not in self.__getattribute__("_tensor_dir")
+        ):
             raise AttributeError(f"{attr} not found")
         _tensor = self.__getattribute__("_tensor")
         return getattr(_tensor, attr)
@@ -447,13 +471,26 @@ class MemmapTensor(object):
     def __getitem__(self, item: INDEX_TYPING) -> torch.Tensor:
         # return self._load_item(memmap_array=self.memmap_array[item])#[item]
         # return self._load_item()[item]
+        if isinstance(item, (NoneType, EllipsisType, int, np.integer, slice)):
+            item = (item,)
         return self._load_item(idx=item)
 
     def __setitem__(self, idx: INDEX_TYPING, value: torch.Tensor):
-        # self.memmap_array[idx] = to_numpy(value)
-        self._load_item()[idx] = value
+        if self.device == torch.device("cpu"):
+            self._load_item()[idx] = value
+        else:
+            if isinstance(idx, torch.Tensor):
+                idx = idx.cpu()
+            elif isinstance(idx, tuple) and any(
+                isinstance(_idx, torch.Tensor) for _idx in idx
+            ):
+                idx = tuple(
+                    _idx.cpu() if isinstance(_idx, torch.Tensor) else _idx
+                    for _idx in idx
+                )
+            self.memmap_array[idx] = to_numpy(value)
 
-    def __setstate__(self, state: dict) -> None:
+    def __setstate__(self, state: Dict[str, Any]) -> None:
         if state["file"] is None:
             # state["_had_ownership"] = state["_had_ownership"]
             # state["_has_ownership"] = delete
@@ -464,7 +501,7 @@ class MemmapTensor(object):
             state["file"] = tmpfile
         self.__dict__.update(state)
 
-    def __getstate__(self) -> dict:
+    def __getstate__(self) -> Dict[str, Any]:
         state = self.__dict__.copy()
         state["file"] = None
         state["_memmap_array"] = None
@@ -529,36 +566,53 @@ class MemmapTensor(object):
         return tuple(self[_idx] for _idx in idx)
 
 
-@implements_for_memmap(torch.stack)
-def stack(
+def _stack(
     list_of_memmap: List[MemmapTensor],
     dim: int,
     out: Optional[Union[torch.Tensor]] = None,
 ) -> torch.Tensor:
     list_of_tensors = [
-        a._tensor if isinstance(a, MemmapTensor) else a for a in list_of_memmap
+        a._tensor if isinstance(a, (MemmapTensor,)) else a for a in list_of_memmap
     ]
-    return torch.stack(list_of_tensors, dim, out=out)
+    if isinstance(out, (MemmapTensor,)):
+        list_of_tensors = [tensor.cpu() for tensor in list_of_tensors]
+        return torch.stack(list_of_tensors, dim, out=out._tensor_from_numpy)
+    else:
+        return torch.stack(list_of_tensors, dim, out=out)
 
 
-@implements_for_memmap(torch.unbind)
-def unbind(memmap: MemmapTensor, dim: int) -> Tuple[torch.Tensor, ...]:
+implements_for_memmap(torch.stack)(_stack)
+
+
+def _unbind(memmap: MemmapTensor, dim: int) -> Tuple[torch.Tensor, ...]:
     return memmap.unbind(dim)
 
 
-@implements_for_memmap(torch.cat)
-def cat(
+implements_for_memmap(torch.unbind)(_unbind)
+
+
+def _tensor(memmap: MemmapTensor) -> torch.Tensor:
+    return memmap._tensor
+
+
+implements_for_memmap(torch.tensor)(_tensor)
+
+
+def _cat(
     list_of_memmap: List[MemmapTensor],
     dim: int,
     out: Optional[Union[torch.Tensor, MemmapTensor]] = None,
 ) -> torch.Tensor:
     list_of_tensors = [
-        a._tensor if isinstance(a, MemmapTensor) else a for a in list_of_memmap
+        a._tensor if isinstance(a, (MemmapTensor,)) else a for a in list_of_memmap
     ]
     print("mm: ", [t.shape for t in list_of_memmap])
     print("tensors: ", [t.shape for t in list_of_tensors])
     print("dim: ", dim, "shape: ", torch.cat(list_of_tensors, dim, out=out).shape)
     return torch.cat(list_of_tensors, dim, out=out)
+
+
+implements_for_memmap(torch.cat)(_cat)
 
 
 def set_transfer_ownership(memmap: MemmapTensor, value: bool = True) -> None:
